requirements.txt
streamlit>=1.30.0
apify-client>=1.6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
anthropic>=0.30.0
